requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
python-telegram-bot==13.15
python-dotenv==1.0.0
schedule==1.2.1
lxml==4.9.3
funpayapi==1.1.0 